% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{buildXYData}
\alias{buildXYData}
\title{Glimma XY Plot}
\usage{
buildXYData(
  table,
  status,
  main,
  display.columns,
  anno,
  counts,
  xlab,
  ylab,
  status.colours,
  groups,
  transform.counts
)
}
\description{
Common processing steps for both MA and XY plots.
}
