% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{extractGroups}
\alias{extractGroups}
\title{extractGroups}
\usage{
extractGroups(x)
}
\arguments{
\item{x}{SummarizedExperiment object}
}
\description{
extract groups vector from colData(x) if it is present; otherwise return
vector of 1s.
}
