% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlimmaV2.R
\name{glimmaMA.default}
\alias{glimmaMA.default}
\title{Glimma MA Plot}
\usage{
\method{glimmaMA}{default}(
  x,
  status = rep(0, nrow(x)),
  main = "MA Plot",
  coef = ncol(x$coefficients),
  p.adj.method = "BH",
  display.columns = NULL,
  anno = NULL,
  xlab = NULL,
  ylab = NULL,
  status.colours = c("dodgerblue", "lightslategray", "firebrick"),
  width = 900,
  height = 570
)
}
\description{
Glimma MA Plot
}
