% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaMA.DGEExact}
\alias{glimmaMA.DGEExact}
\title{Glimma MA Plot}
\usage{
\method{glimmaMA}{DGEExact}(
  x,
  status = edgeR::decideTestsDGE(x),
  main = paste(x$comparison[2], "vs", x$comparison[1]),
  p.adj.method = "BH",
  display.columns = NULL,
  anno = NULL,
  groups = NULL,
  counts = NULL,
  xlab = NULL,
  ylab = NULL,
  status.colours = NULL,
  transform.counts = FALSE,
  width = 920,
  height = 920
)
}
\description{
Glimma MA Plot
}
