% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaMA.DGELRT}
\alias{glimmaMA.DGELRT}
\title{Glimma MA Plot}
\usage{
\method{glimmaMA}{DGELRT}(
  x,
  status = NULL,
  main = paste(x$comparison[2], "vs", x$comparison[1]),
  p.adj.method = "BH",
  display.columns = NULL,
  anno = NULL,
  counts = NULL,
  xlab = NULL,
  ylab = NULL,
  status.colours = c("dodgerblue", "lightslategray", "firebrick"),
  width = 900,
  height = 570
)
}
\description{
Glimma MA Plot
}
