% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaMA.MArrayLM}
\alias{glimmaMA.MArrayLM}
\title{Glimma MA Plot}
\usage{
\method{glimmaMA}{MArrayLM}(
  x,
  status = rep(0, nrow(x)),
  coef = ncol(x$coefficients),
  main = colnames(x)[coef],
  p.adj.method = "BH",
  display.columns = NULL,
  anno = NULL,
  counts = NULL,
  xlab = NULL,
  ylab = NULL,
  status.colours = c("dodgerblue", "lightslategray", "firebrick"),
  width = 900,
  height = 570
)
}
\description{
Glimma MA Plot
}
