% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaMDS.R
\name{glimmaMDS.default}
\alias{glimmaMDS.default}
\title{Glimma MDS Plot}
\usage{
\method{glimmaMDS}{default}(
  x,
  top = 500,
  labels = as.character(seq_len(ncol(x))),
  groups = as.character(rep(1, ncol(x))),
  gene.selection = c("pairwise", "common"),
  continuous.colour = FALSE,
  save = FALSE,
  html = NULL,
  width = 900,
  height = 500
)
}
\arguments{
\item{x}{the matrix containing the gene expressions.}

\item{top}{the number of top most variable genes to use.}

\item{labels}{the labels for each sample.}

\item{groups}{the experimental group to which samples belong.}

\item{gene.selection}{"pairwise" if most variable genes are to be chosen for each pair of samples or
"common" to select the same genes for all comparisons.}

\item{continuous.colour}{if \code{TRUE}, colour is displayed using continuous columns in groups; otherwise,
colour is displayed using discrete factors.}

\item{html}{name of HTML file (including extension) to export widget into rather than displaying the widget; \code{NULL} by default.}

\item{width}{custom widget width in pixels}

\item{height}{custom widget height in pixels}
}
\value{
htmlwidget object or \code{NULL} if \code{html} argument is specified.
}
\description{
Draws a two-panel interactive MDS plot.
}
\details{
The left plot shows two MDS dimensions, with annotations displayed on hover.
The right panel contains a bar plot of the eigenvalues of each dimension.
The controls beneath the plots can be used to change the dimensions being displayed.
}
\seealso{
\code{\link{glimmaMDS}}, \code{\link{glimmaMDS.DGEList}}, \code{\link{glimmaMDS.DESeqDataSet}}
}
