% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaMDS.R
\name{glimmaMDS.DESeqDataSet}
\alias{glimmaMDS.DESeqDataSet}
\title{Glimma MDS Plot}
\usage{
\method{glimmaMDS}{DESeqDataSet}(
  x,
  top = 500,
  labels = NULL,
  groups = NULL,
  gene.selection = c("pairwise", "common"),
  prior.count = 0.25,
  continuous.colour = FALSE,
  width = 900,
  height = 570
)
}
\arguments{
\item{x}{the matrix containing the gene expressions.}

\item{top}{the number of top most variable genes to use.}

\item{labels}{the labels for each sample.}

\item{groups}{the experimental group to which samples belong.}

\item{gene.selection}{"pairwise" if most variable genes are to be chosen for each pair of samples or
"common" to select the same genes for all comparisons.}

\item{prior.count}{average count to be added to each observation to avoid taking log of zero.}

\item{continuous.colour}{if TRUE, colour is displayed using continuous columns in groups; if false,
colour is displayed using discrete columns.}

\item{width}{custom widget width in pixels}

\item{height}{custom widget height in pixels}
}
\description{
Draws a two-panel interactive MDS plot using a DESeqDataset x. 
By default, extracts groups as \code{colData(x)} and extracts labels as \code{rownames(colData(x))}.
}
\seealso{
\code{\link{glimmaMDS.default}}, \code{\link{glimmaMDS.DGEList}}
}
