% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlimmaV2.R
\name{GlimmaV2}
\alias{GlimmaV2}
\title{GlimmaV2 HTMLWidget Wrapper}
\usage{
GlimmaV2(xData, width = 900, height = 570, elementId = NULL, ...)
}
\description{
Passes packaged data to JS interface for rendering.
}
