% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaVolcano.R
\name{glimmaVolcano.MArrayLM}
\alias{glimmaVolcano.MArrayLM}
\title{Glimma Volcano Plot}
\usage{
\method{glimmaVolcano}{MArrayLM}(
  x,
  dge = NULL,
  status = limma::decideTests(x),
  coef = ncol(x$coefficients),
  main = colnames(x)[coef],
  p.adj.method = "BH",
  display.columns = NULL,
  anno = x$genes,
  groups = dge$samples$group,
  counts = dge$counts,
  xlab = "logFC",
  ylab = "negLog10PValue",
  status.colours = c("dodgerblue", "silver", "firebrick"),
  transform.counts = FALSE,
  html = NULL,
  width = 920,
  height = 920
)
}
\arguments{
\item{x}{\code{MArrayLM} object from which summary statistics are extracted from to create 
summary (left) plot.}

\item{dge}{\code{DGEList} object with \code{nrow(x)} rows from which expression values are 
extracted from to create expression (right) plot. Gene counts are taken from \code{dge$counts} 
and sample groups from \code{dge$samples$group}.}

\item{status}{vector of length \code{nrow(x)} indicating the status of each gene. 
By default genes in the summary plot are coloured based on its differential expression status 
using an adjusted p-value cutoff of 5\% by calling the \code{limma::decideTests} function, where
the value of -1 marks down-regulated genes, 0 marks genes with no expression difference, and 
1 marks up-regulated genes.}

\item{coef}{integer indicating the column in \code{x} from the summary plot is created.}

\item{main}{character string for the main title of summary plot.}

\item{p.adj.method}{character string specifying p-value adjustment method.}

\item{display.columns}{character vector containing names of columns from \code{anno} from
which to display in mouseover tooltips and table.}

\item{anno}{dataframe with \code{nrow(x)} rows containing gene annotations.}

\item{groups}{vector of length \code{ncol(dge)} representing categorisation of samples in 
expression plot.}

\item{counts}{numeric matrix with \code{nrow(x)} rows containing gene expression values. 
This can be used to replace raw gene counts from \code{dge$counts} with transformed counts 
e.g. logCPM or logRPKM values.}

\item{xlab}{character string for the x-axis label of summary plot.}

\item{ylab}{character string for the y-axis label of summary plot.}

\item{status.colours}{vector of length 3 containing valid CSS strings for colours associated
with \code{status}  in the order of -1, 0 and 1.}

\item{transform.counts}{TRUE if counts should be log-cpm transformed, defaults to FALSE.}

\item{html}{character string for naming HTML file for exportation of widget. The extension 
should be included in the file name e.g. "file.html".}

\item{width}{numeric value indicating width of widget in pixels.}

\item{height}{numeric value indicating width of height in pixels.}
}
\value{
htmlwidget object or \code{NULL} if \code{html} argument is specified.
}
\description{
Draws a two-panel interactive volcano plot from an MArrayLM object.
}
\details{
The left plot shows the log-fold-change vs -log10(pvalue).
The right plot shows the expression levels of a particular gene of each sample.
Clicking on genes in the plot brings up the corresponding genes in the table.
Selecting rows in the table will highlight the corresponding genes in the volcano plot.
Expression values for a gene can be found by hovering over a sample in the right plot.
}
\seealso{
\code{\link{glimmaVolcano}}, \code{\link{glimmaVolcano.DGEExact}}, \code{\link{glimmaVolcano.DGELRT}}, \code{\link{glimmaVolcano.DESeqDataSet}}
}
