% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaVolcano.R
\name{glimmaVolcano}
\alias{glimmaVolcano}
\title{Glimma Volcano Plot}
\usage{
glimmaVolcano(x, ...)
}
\arguments{
\item{x}{the DE object to plot.}

\item{...}{additional arguments affecting the plots produced. See specific methods for detailed arguments.}
}
\value{
htmlwidget object.
}
\description{
Draws a two-panel interactive MA plot.
}
\details{
The left plot shows the log-fold-change vs -log10(pvalue).
The right plot shows the expression levels of a particular gene of each sample.
Clicking on genes in the plot brings up the corresponding genes in the table.
Selecting rows in the table will highlight the corresponding genes in the volcano plot.
Expression values for a gene can be found by hovering over a sample in the right plot.
}
\seealso{
\code{\link{glimmaVolcano.MArrayLM}}, \code{\link{glimmaVolcano.DGEExact}}, \code{\link{glimmaVolcano.DGELRT}}, \code{\link{glimmaVolcano.DESeqDataSet}}
}
