% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaXY.default}
\alias{glimmaXY.default}
\title{Glimma XY Plot}
\usage{
\method{glimmaXY}{default}(
  x,
  y,
  xlab = "x",
  ylab = "y",
  status = rep(0, length(x)),
  main = "XY Plot",
  display.columns = NULL,
  anno = NULL,
  groups = NULL,
  counts = NULL,
  status.colours = NULL,
  transform.counts = FALSE,
  width = 920,
  height = 920
)
}
\description{
Glimma XY Plot
}
