% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaXY}
\alias{glimmaXY}
\title{Glimma XY Plot}
\usage{
glimmaXY(x, ...)
}
\arguments{
\item{x}{the DE object to plot.}

\item{...}{additional arguments affecting the plots produced. See specific methods for detailed arguments.}
}
\value{
htmlwidget object.
}
\description{
Draws a two-panel interactive XY scatter plot.
}
\details{
The left plot shows the x and y values specified.
The right plot shows the expression levels of a particular gene of each sample.
Clicking on genes in the plot brings up the corresponding genes in the table.
Selecting rows in the table will highlight the corresponding genes in the XY plot.
Expression values for a gene can be found by hovering over a sample in the RHS plot.
}
