% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaXYWidget}
\alias{glimmaXYWidget}
\title{GlimmaXY HTMLWidget Wrapper}
\usage{
glimmaXYWidget(xData, width = 900, height = 570, elementId = NULL, ...)
}
\description{
Passes packaged data to JS interface for rendering.
}
