% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1-glScatter.R
\name{glScatter}
\alias{glScatter}
\title{Glimma Scatter Plot}
\usage{
glScatter(x, ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{...}{additional arguments depending on input object type.}
}
\value{
A chart object containing the information to create an interactive scatter plot.
}
\description{
Create an interactive scatter plot object
}
\author{
Shian Su
}
\keyword{internal}
