% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{buildXYData}
\alias{buildXYData}
\title{Glimma XY Plot}
\usage{
buildXYData(
  table,
  status,
  main,
  display.columns,
  anno,
  counts,
  xlab,
  ylab,
  status.colours,
  groups,
  transform.counts
)
}
\arguments{
\item{table}{dataframe containing xlab and ylab columns for plotting.}

\item{status}{vector giving the control status of each row in the object.}

\item{main}{Left plot title.}

\item{display.columns}{character vector containing names of columns to display in mouseover tooltips and table.}

\item{anno}{a dataframe containing gene annotations}

\item{counts}{the matrix of expression values, with samples in columns. Will override dge$counts if dge is provided too.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{status.colours}{vector of three valid CSS strings representing colours for genes with status [-1, 0 and 1] respectively.}

\item{groups}{vector/factor representing the experimental group for each sample. Will override dge$samples$group if dge is provided too.}

\item{transform.counts}{TRUE if counts should be log-cpm transformed, defaults to FALSE.}
}
\description{
Common processing steps for both MA and XY plots.
}
