% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{extractGroups}
\alias{extractGroups}
\title{extractGroups}
\usage{
extractGroups(cdata)
}
\arguments{
\item{cdata}{SummarizedExperiment column data matrix}
}
\value{
groups column of data if present, otherwise 1
}
\description{
Extracts the column named \code{group} from column data matrix
of a SummarizedExperiment object if it is present. Otherwise return a
vector of 1s.
}
\keyword{internal}
