% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaMA.DESeqDataSet}
\alias{glimmaMA.DESeqDataSet}
\title{Glimma MA Plot}
\usage{
\method{glimmaMA}{DESeqDataSet}(
  x,
  status = NULL,
  main = "MA Plot",
  display.columns = NULL,
  anno = NULL,
  groups = NULL,
  counts = DESeq2::counts(x),
  xlab = NULL,
  ylab = NULL,
  status.colours = NULL,
  transform.counts = FALSE,
  width = 920,
  height = 920
)
}
\description{
Glimma MA Plot
}
