% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaMA.R
\name{glimmaMA.DESeqDataSet}
\alias{glimmaMA.DESeqDataSet}
\title{Glimma MA Plot}
\usage{
\method{glimmaMA}{DESeqDataSet}(
  x,
  status = NULL,
  main = "MA Plot",
  display.columns = NULL,
  anno = NULL,
  groups = extractGroups(colData(x)),
  counts = DESeq2::counts(x),
  xlab = "logCPM",
  ylab = "logFC",
  status.colours = c("dodgerblue", "silver", "firebrick"),
  transform.counts = FALSE,
  save = FALSE,
  filename = "glimmaMA.html",
  width = 920,
  height = 920
)
}
\arguments{
\item{x}{the MArrayLM object to plot.}

\item{status}{vector giving the control status of each row in the object.}

\item{main}{Left plot title.}

\item{display.columns}{character vector containing names of columns to display in mouseover tooltips and table.}

\item{anno}{a dataframe containing gene annotations.}

\item{groups}{vector/factor representing the experimental group for each sample; see \code{\link{extractGroups}} for default value.}

\item{counts}{the matrix of expression values, with samples in columns.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{status.colours}{vector of three valid CSS strings representing colours for genes with status [-1, 0 and 1] respectively.}

\item{transform.counts}{TRUE if counts should be log-cpm transformed, defaults to FALSE.}

\item{save}{if \code{TRUE}, widget will be exported to standalone HTML file rather than being displayed.}

\item{filename}{name of the standalone HTML file created if \code{save} is \code{TRUE}.}

\item{width}{width of the weidget in pixels.}

\item{height}{height of the widget in pixels.}
}
\value{
htmlwidget object or \code{NULL} if \code{save} argument is \code{TRUE}.
}
\description{
Draws a two-panel interactive MA plot from an DESeqDataSet object.
}
\details{
The left plot shows the log-fold-change vs average expression.
The right plot shows the expression levels of a particular gene of each sample.
Clicking on genes in the plot brings up the corresponding genes in the table.
Selecting rows in the table will highlight the corresponding genes in the MA plot.
Expression values for a gene can be found by hovering over a sample in the right plot.
}
\seealso{
\code{\link{glimmaMA}}, \code{\link{glimmaMA.MArrayLM}}, \code{\link{glimmaMA.DGEExact}}, \code{\link{glimmaMA.DGELRT}}
}
