% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaMA.MArrayLM}
\alias{glimmaMA.MArrayLM}
\title{Glimma MA Plot}
\usage{
\method{glimmaMA}{MArrayLM}(
  x,
  status = limma::decideTests(x),
  coef = ncol(x$coefficients),
  main = colnames(x)[coef],
  p.adj.method = "BH",
  display.columns = NULL,
  anno = NULL,
  groups = NULL,
  counts = NULL,
  xlab = NULL,
  ylab = NULL,
  status.colours = NULL,
  transform.counts = FALSE,
  width = 920,
  height = 920
)
}
\description{
Glimma MA Plot
}
