% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaMA.MArrayLM}
\alias{glimmaMA.MArrayLM}
\title{Glimma MA Plot}
\usage{
\method{glimmaMA}{MArrayLM}(
  x,
  dge = NULL,
  status = limma::decideTests(x),
  coef = ncol(x$coefficients),
  main = colnames(x)[coef],
  p.adj.method = "BH",
  display.columns = NULL,
  anno = x$genes,
  groups = NULL,
  counts = NULL,
  xlab = "logCPM",
  ylab = "logFC",
  status.colours = NULL,
  transform.counts = FALSE,
  width = 920,
  height = 920
)
}
\arguments{
\item{x}{the MArrayLM object to plot.}

\item{dge}{DGEList object used to generate the expression plot. Gene counts are extracted from \code{dge$counts} and
sample groups from \code{dge$samples$group}.}

\item{status}{vector giving the control status of each row in the object.}

\item{coef}{integer or logical indexing vector indicating which column of object to plot.}

\item{main}{Left plot title.}

\item{p.adj.method}{character vector indicating multiple testing correction method. See p.adjust for available methods. (defaults to "BH")}

\item{display.columns}{character vector containing names of columns to display in mouseover tooltips and table.}

\item{anno}{a dataframe containing gene annotations}

\item{groups}{vector/factor representing the experimental group for each sample. Will override dge$samples$group if dge is provided too.}

\item{counts}{the matrix of expression values, with samples in columns. Will override dge$counts if dge is provided too.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{status.colours}{vector of three valid CSS strings representing colours for genes with status [-1, 0 and 1] respectively.}

\item{transform.counts}{TRUE if counts should be log-cpm transformed, defaults to FALSE.}

\item{width}{width of the weidget in pixels.}

\item{height}{height of the widget in pixels.}
}
\value{
htmlwidget object.
}
\description{
Draws a two-panel interactive MA plot from an MArrayLM object.
}
\details{
The left plot shows the log-fold-change vs average expression.
The right plot shows the expression levels of a particular gene of each sample.
Clicking on genes in the plot brings up the corresponding genes in the table.
Selecting rows in the table will highlight the corresponding genes in the MA plot.
Expression values for a gene can be found by hovering over a sample in the right plot.
}
\seealso{
\code{\link{glimmaMA.DGEExact}}, \code{\link{glimmaMA.DGELRT}}, \code{\link{glimmaMA.DESeqDataSet}}
}
