% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaMA.R
\name{glimmaMA}
\alias{glimmaMA}
\title{Glimma MA Plot}
\usage{
glimmaMA(x, ...)
}
\arguments{
\item{x}{the DE object to plot.}

\item{...}{additional arguments affecting the plots produced. See specific methods for detailed arguments.}
}
\value{
htmlwidget object or \code{NULL} if \code{save} argument is \code{TRUE}.
}
\description{
Generic function for drawing a two-panel interactive MA plot. 
The function invokes the following methods which depend on the class of the first argument:
\itemize{
  \item \code{\link{glimmaMA.MArrayLM}} for limma analysis
  \item \code{\link{glimmaMA.DGEExact}} for edgeR analysis, produced from \code{\link{exactTest}}
  \item \code{\link{glimmaMA.DGELRT}} for edgeR analysis, produced from \code{\link{glmLRT}}
  \item \code{\link{glimmaMA.DESeqDataSet}} for DESeq2 analysis }
}
\details{
The left plot shows the log-fold-change vs average expression.
The right plot shows the expression levels of a particular gene of each sample.
Clicking on genes in the plot brings up the corresponding genes in the table.
Selecting rows in the table will highlight the corresponding genes in the MA plot.
Expression values for a gene can be found by hovering over a sample in the right plot.
}
