% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaMDS.R
\name{glimmaMDS.default}
\alias{glimmaMDS.default}
\title{Glimma MDS Plot}
\usage{
\method{glimmaMDS}{default}(
  x,
  top = 500,
  labels = as.character(seq_len(ncol(x))),
  groups = as.character(rep(1, ncol(x))),
  gene.selection = c("pairwise", "common"),
  continuous.colour = FALSE,
  width = 900,
  height = 570
)
}
\arguments{
\item{x}{the matrix containing the gene expressions.}

\item{top}{the number of top most variable genes to use.}

\item{labels}{the labels for each sample.}

\item{groups}{the experimental group to which samples belong.}

\item{gene.selection}{"pairwise" if most variable genes are to be chosen for each pair of samples or
"common" to select the same genes for all comparisons.}

\item{continuous.colour}{if TRUE, colour is displayed using continuous columns in groups; if false,
colour is displayed using discrete columns.}

\item{width}{custom widget width in pixels}

\item{height}{custom widget height in pixels}
}
\description{
Draws a two-panel interactive MDS plot. The left panel contains the plot between
two MDS dimensions, with annotations displayed on hover. The right panel contains a bar plot of
the eigenvalues of each dimension. The controls below can be used to change the dimensions being
displayed.
}
\seealso{
\code{\link{glimmaMDS.DGEList}}, \code{\link{glimmaMDS.DESeqDataSet}}
}
