% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaMDS.R
\name{glimmaMDS.DESeqDataSet}
\alias{glimmaMDS.DESeqDataSet}
\title{Glimma MDS Plot}
\usage{
\method{glimmaMDS}{DESeqDataSet}(
  x,
  top = 500,
  labels = rownames(SummarizedExperiment::colData(x)),
  groups = as.data.frame(SummarizedExperiment::colData(x)),
  gene.selection = c("pairwise", "common"),
  prior.count = 0.25,
  continuous.colour = FALSE,
  html = NULL,
  width = 900,
  height = 500
)
}
\arguments{
\item{x}{the matrix containing the gene expressions.}

\item{top}{the number of top most variable genes to use.}

\item{labels}{the labels for each sample.}

\item{groups}{the experimental group to which samples belong.}

\item{gene.selection}{"pairwise" if most variable genes are to be chosen for each pair of samples or
"common" to select the same genes for all comparisons.}

\item{prior.count}{average count to be added to each observation to avoid taking log of zero.}

\item{continuous.colour}{if \code{TRUE}, colour is displayed using continuous columns in groups; otherwise,
colour is displayed using discrete factors.}

\item{html}{name of HTML file (including extension) to export widget into rather than displaying the widget; \code{NULL} by default.}

\item{width}{custom widget width in pixels}

\item{height}{custom widget height in pixels}
}
\value{
htmlwidget object or \code{NULL} if \code{save} argument is \code{TRUE}.
}
\description{
Draws a two-panel interactive MDS plot using a DESeqDataset x.
}
\details{
The left plot shows two MDS dimensions, with annotations displayed on hover.
The right panel contains a bar plot of the eigenvalues of each dimension.
The controls beneath the plots can be used to change the dimensions being displayed.
}
\seealso{
\code{\link{glimmaMDS}}, \code{\link{glimmaMDS.default}}, \code{\link{glimmaMDS.DGEList}}
}
