% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlimmaV2.R
\name{glimmaMDS}
\alias{glimmaMDS}
\title{Glimma MDS Plot}
\usage{
glimmaMDS(x, ...)
}
\arguments{
\item{x}{the matrix containing the gene expressions.}

\item{...}{the additional arguments.}
}
\description{
Draws a two-panel interactive MDS plot in an html page. The left panel contains the plot between
two MDS dimensions, with annotations displayed on hover. The right panel contains a bar plot of
the eigenvalues of each dimension. The controls below can be used to change the dimensions being
displayed.
}
\seealso{
\code{\link{glimmaMDS.default}}, \code{\link{glimmaMDS.DGEList}}, \code{\link{glimmaMDS.DESeqDataSet}}
}
