% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaMDS.R
\name{glimmaMDS}
\alias{glimmaMDS}
\title{Glimma MDS Plot}
\usage{
glimmaMDS(x, ...)
}
\arguments{
\item{x}{the matrix containing the gene expressions.}

\item{...}{the additional arguments affecting the plot produced. See specific methods for detailed arguments.}
}
\value{
htmlwidget object or \code{NULL} if \code{html} argument is specified.
}
\description{
Generic function for drawing a two-panel interactive MDS plot.
The function invokes the following methods which depend on the class of the first argument:
\itemize{
  \item \code{\link{glimmaMDS.DGEList}} for edgeR analysis
  \item \code{\link{glimmaMDS.DESeqDataSet}} for DESeq2 analysis
  \item \code{\link{glimmaMDS.default}} for all other object types }
}
\details{
The left plot shows two MDS dimensions, with annotations displayed on hover.
The right panel contains a bar plot of the eigenvalues of each dimension.
The controls beneath the plots can be used to change the dimensions being displayed.
}
