% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlimmaV2.R
\name{glimmaXY.default}
\alias{glimmaXY.default}
\title{Glimma XY Plot}
\usage{
\method{glimmaXY}{default}(
  x,
  y,
  xlab = "x",
  ylab = "y",
  status = rep(0, length(x)),
  main = "XY Plot",
  display.columns = NULL,
  anno = NULL,
  counts = NULL,
  status.colours = c("dodgerblue", "lightslategray", "firebrick"),
  width = 900,
  height = 570
)
}
\description{
Glimma XY Plot
}
