% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaXY}
\alias{glimmaXY}
\title{Glimma XY Plot}
\usage{
glimmaXY(
  x,
  y,
  xlab = "x",
  ylab = "y",
  status = rep(0, length(x)),
  main = "XY Plot",
  display.columns = NULL,
  anno = NULL,
  groups = NULL,
  counts = NULL,
  status.colours = c("dodgerblue", "silver", "firebrick"),
  transform.counts = FALSE,
  html = NULL,
  width = 920,
  height = 920
)
}
\arguments{
\item{x}{numeric vector of values to plot on the x-axis of the summary plot.}

\item{y}{numeric vector of values to plot on the y-axis of the summary plot.}

\item{xlab}{character string for the x-axis label of summary plot.}

\item{ylab}{character string for the y-axis label of summary plot.}

\item{status}{vector of length \code{nrow(x)} indicating the status of each gene. 
By default genes in the summary plot are coloured based on its differential expression status 
using an adjusted p-value cutoff of 5\% by calling the \code{limma::decideTests} function, where
the value of -1 marks down-regulated genes, 0 marks genes with no expression difference, and 
1 marks up-regulated genes.}

\item{main}{character string for the main title of summary plot.}

\item{display.columns}{character vector containing names of columns from \code{anno} from
which to display in mouseover tooltips and table.}

\item{anno}{dataframe with \code{nrow(x)} rows containing gene annotations.}

\item{groups}{vector of length \code{ncol(dge)} representing categorisation of samples in 
expression plot.}

\item{counts}{numeric matrix with \code{nrow(x)} rows containing gene expression values. 
This can be used to replace raw gene counts from \code{dge$counts} with transformed counts 
e.g. logCPM or logRPKM values.}

\item{status.colours}{vector of length 3 containing valid CSS strings for colours associated
with \code{status}  in the order of -1, 0 and 1.}

\item{transform.counts}{TRUE if counts should be log-cpm transformed, defaults to FALSE.}

\item{html}{character string for naming HTML file for exportation of widget. The extension 
should be included in the file name e.g. "file.html".}

\item{width}{numeric value indicating width of widget in pixels.}

\item{height}{numeric value indicating width of height in pixels.}
}
\value{
htmlwidget object or \code{NULL} if \code{html} argument is specified.
}
\description{
Draws a two-panel interactive XY scatter plot.
}
\details{
The left plot shows the x and y values specified.
The right plot shows the expression levels of a particular gene of each sample.
Clicking on genes in the plot brings up the corresponding genes in the table.
Selecting rows in the table will highlight the corresponding genes in the XY plot.
Expression values for a gene can be found by hovering over a sample in the right plot.
}
