% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaXYWidget}
\alias{glimmaXYWidget}
\title{GlimmaXY HTMLWidget Wrapper}
\usage{
glimmaXYWidget(xData, width, height, save, filename)
}
\arguments{
\item{xData}{packaged data object returned from buildXYData()}

\item{width}{htmlwidget element width in pixels}

\item{height}{htmlwidget element height in pixels}

\item{save}{if \code{TRUE}, widget will be exported to standalone HTML file rather than being displayed.}

\item{filename}{name of the standalone HTML file created if \code{save} is \code{TRUE}.}
}
\description{
Passes packaged data to JS interface for rendering.
}
