% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaXYWidget}
\alias{glimmaXYWidget}
\title{GlimmaXY HTMLWidget Wrapper}
\usage{
glimmaXYWidget(xData, width, height, save, filename)
}
\arguments{
\item{xData}{packaged data object returned from buildXYData()}

\item{width}{htmlwidget element width in pixels}

\item{height}{htmlwidget element height in pixels}

\item{elementId}{ID attribute for htmlwidget}
}
\description{
Passes packaged data to JS interface for rendering.
}
