% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmaXY.R
\name{glimmaXYWidget}
\alias{glimmaXYWidget}
\title{GlimmaXY HTMLWidget Wrapper}
\usage{
glimmaXYWidget(xData, width, height, html)
}
\arguments{
\item{xData}{packaged data object returned from buildXYData()}

\item{width}{htmlwidget element width in pixels}

\item{height}{htmlwidget element height in pixels}

\item{html}{name of HTML file (including extension) to export widget into rather than displaying the widget; \code{NULL} by default.}
}
\value{
htmlwidget object for XY plot internal use
}
\description{
Passes packaged data to JS interface for rendering.
}
\keyword{internal}
